feat: remove gosling.js/embed and recommend esm.sh for Observable
