+++ package.json
@@ -15,14 +15,12 @@
         "dist"
     ],
     "exports": {
-        ".": "./dist/gosling.es.js",
-        "./embed": "./dist/embed/index.js"
+        ".": "./dist/gosling.es.js"
     },
     "scripts": {
         "start": "vite --mode editor",
-        "start-embed": "vite",
         "build": "run-s build-clear build-types build-lib",
-        "build-lib": "vite build --mode lib && node scripts/build-umd && node scripts/build-embed",
+        "build-lib": "vite build --mode lib && node scripts/build-umd",
         "build-types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
         "build-editor": "node --max_old_space_size=4096 ./node_modules/vite/bin/vite.js build",
         "build-clear": "rm -rf ./dist",
