+++ tsconfig.json
@@ -33,8 +33,7 @@
   },
   "files": [
     "src/index.ts",
-    "editor/index.tsx",
-    "embed/index.ts"
+    "editor/index.tsx"
   ],
   "include": ["src", "src/**/*.d.ts", "editor"],
   "exclude": [
